+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ## Unreleased
 * Add AuthorizationAdjustment class and `authorizationAdjustments` to Transaction
-
-## Unreleased
+* Add iDEAL webhook support
 * Add `IDEAL_PAYMENT` to `PaymentInstrumentType`
 * Create Braintree class to be PSR compliant